feat/refactor (Landing to 'home' page, moved Sign In comp)
